Move getIdFromUrl into utils
